Fix tests errors due to moving processing to processing_time_series in mango_time_series.
Run black on all files.
