+++ mango/mango/clients/aemet.py
@@ -255,7 +255,10 @@ class AEMETClient(RESTClient):
         ]
         objective_lat_long = (lat, long)
         distances = [
-            (station_indicative, haversine(point1=objective_lat_long, point2=station_coords))
+            (
+                station_indicative,
+                haversine(point1=objective_lat_long, point2=station_coords),
+            )
             for station_indicative, station_coords in lat_long_list
         ]
         # Sort by distance
