+++ mango/mango/shap_analysis/const.py
@@ -1,4 +1,5 @@
 """ This file contains the constants used in the shap_analysis module. """
+
 TREE_EXPLAINER_MODELS = [
     "XGBRegressor",
     "LGBMClassifier",
