+++ mango/mango/shared/__init__.py
@@ -1,4 +1,10 @@
 from .const import ARCGIS_TOKEN_URL
 from .decorators import validate_args, pydantic_validation
-from .exceptions import InvalidCredentials, ValidationError, JobError, BadResponse, ApiKeyError
+from .exceptions import (
+    InvalidCredentials,
+    ValidationError,
+    JobError,
+    BadResponse,
+    ApiKeyError,
+)
 from .spatial import haversine
