+++ mango/mango/tests/initialization_library/test_init.py
@@ -6,7 +6,7 @@ from mango.tests.const import normalize_path
 
 class InitializeTest(TestCase):
     def test_version_matches(self):
-        with open(normalize_path(os.path.join("..","..","pyproject.toml")),"r") as f:
+        with open(normalize_path(os.path.join("..", "..", "pyproject.toml")), "r") as f:
             pyproject = f.read()
-        version = pyproject.split("version = ")[1].split("\n")[0].replace('"','')
-        self.assertEqual(mango.__version__,version)
+        version = pyproject.split("version = ")[1].split("\n")[0].replace('"', "")
+        self.assertEqual(mango.__version__, version)
