+++ mango/mango/tests/shared_module/test_requirement_check.py
@@ -5,7 +5,8 @@ from pkg_resources import DistributionNotFound
 
 from mango.shared.requirement_check import check_dependencies
 
-def raise_exception(*args,**kwargs):
+
+def raise_exception(*args, **kwargs):
     raise DistributionNotFound()
 
 
@@ -19,7 +20,7 @@ class TestRequirementCheck(TestCase):
             "option_1 = ['mocklib<=1.12']\n"
             "option_2 = ['impossible_package_name<=1.12']"
         )
-        with open("mockup.toml","w") as toml_file:
+        with open("mockup.toml", "w") as toml_file:
             toml_file.write(toml_data)
 
     @classmethod
@@ -27,18 +28,26 @@ class TestRequirementCheck(TestCase):
         os.remove("mockup.toml")
 
     @mock.patch("mango.shared.requirement_check.pkg_resources")
-    def test_check_dependencies(self,mock_pkg_resources):
+    def test_check_dependencies(self, mock_pkg_resources):
         mock_pkg_resources.require.value = None
         flag = check_dependencies("option_1", pyproject_path="mockup.toml")
         self.assertTrue(flag)
-    @mock.patch("mango.shared.requirement_check.pkg_resources",autospec=True)
+
+    @mock.patch("mango.shared.requirement_check.pkg_resources", autospec=True)
     def test_missing_dependency(self, mock_pkg_resources):
         mock_pkg_resources.require.side_effect = raise_exception
-        flag = check_dependencies("option_2",pyproject_path="mockup.toml")
+        flag = check_dependencies("option_2", pyproject_path="mockup.toml")
         self.assertFalse(flag)
 
     def test_invalid_option_name(self):
-        self.assertRaises(KeyError, check_dependencies, "invalid_option",pyproject_path="mockup.toml")
+        self.assertRaises(
+            KeyError, check_dependencies, "invalid_option", pyproject_path="mockup.toml"
+        )
 
     def test_invalid_path(self):
-        self.assertRaises(FileNotFoundError,check_dependencies,"some_option",pyproject_path="invalid_path.toml")
+        self.assertRaises(
+            FileNotFoundError,
+            check_dependencies,
+            "some_option",
+            pyproject_path="invalid_path.toml",
+        )
